Switch to mclo.gs crash report upload, replace GTNH-specific code with configuration, and fix stacktrace deobfuscator (#11)
* Update buildscript

* spotlessApply

* Switch from paste.ubuntu.com to mclo.gs

paste.ubuntu.com requires a login for viewing the paste and that is
annoying. mclo.gs is made for Minecraft logs. It does some special
highlighting and censors certain information that is not relevant.
Drawback is that logs are only valid for 90 days after last viewing.

* Replace GTNH-specific code with configuration parameters

* Update MCP link for stacktrace deobfuscation

* Make stacktrace deobfuscation actually functional

* Create HttpUtils#createConnection to reuse HTTP connection parameters in various parts of the mod

* Address review feedback

* Change deobfuscation mapping source

This removes the necessity to extract a zip archive.